requirements.txt
Flask==3.1.1
gevent==25.5.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1
email-validator==2.1.0.post1 